OSSM-3423: Fix TestSecureGateways on IPv6 (#574)
TestSecureGateways uses image istio/examples-helloworld-v1:latest that
was built 6 years ago and does not support IPv6. I replaced that image
with quay.io/jewertow/examples-helloworld-v1:latest that is built from
the recent code implementation supporting IPv6. This is a temporary
change and we will have to replace this image once the community image
will be available.

Signed-off-by: Jacek Ewertowski <jewertow@redhat.com>